fix: changed request comment part + pop up
